refactor: UserService에서 인증하는 로직 분리
