+++ src/main/java/cart/controller/auth/interceptor/AuthInterceptor.java
@@ -1,5 +1,6 @@
 package cart.controller.auth.interceptor;
 
+import cart.controller.auth.Authorization;
 import cart.controller.auth.dto.LoginUser;
 import cart.controller.auth.util.BasicAuthorizationExtractor;
 import cart.exception.UnauthorizedException;
@@ -13,16 +14,16 @@ import javax.servlet.http.HttpServletResponse;
 @Component
 public class AuthInterceptor implements HandlerInterceptor {
 
-    private final UserService userService;
+    private final Authorization authorization;
 
-    public AuthInterceptor(final UserService userService) {
-        this.userService = userService;
+    public AuthInterceptor(final Authorization authorization) {
+        this.authorization = authorization;
     }
 
     @Override
     public boolean preHandle(final HttpServletRequest request, final HttpServletResponse response, final Object handler) throws Exception {
         LoginUser loginUser = BasicAuthorizationExtractor.extract(request.getHeader("authorization"));
-        if (userService.checkLogin(loginUser.getEmail(), loginUser.getPassword())) {
+        if (authorization.checkLogin(loginUser.getEmail(), loginUser.getPassword())) {
             return HandlerInterceptor.super.preHandle(request, response, handler);
         }
         throw new UnauthorizedException("로그인 정보가 잘못되었습니다.");
