+++ src/main/java/cart/service/UserService.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -25,15 +24,5 @@ public class UserService {
         return allUser.stream()
                 .map(UserResponse::from)
                 .collect(Collectors.toList());
-    }
-
-    @Transactional(readOnly = true)
-    public boolean checkLogin(final String email, final String password) {
-        Optional<User> findUserByEmail = userDao.findBy(email);
-        if (findUserByEmail.isEmpty()) {
-            return false;
-        }
-        User user = findUserByEmail.orElseThrow(() -> new IllegalArgumentException("존재하지 않는 메일입니다."));
-        return user.getPassword().equals(password);
     }
 }
