+++ src/test/java/cart/service/UserServiceTest.java
@@ -14,7 +14,6 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.List;
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -38,46 +37,5 @@ public class UserServiceTest {
         List<UserResponse> userResponses = userService.loadAllUser();
         //then
         assertThat(userResponses).hasSize(3);
-    }
-
-    @DisplayName("이메일과 비밀번호가 일치하면 true를 반환한다.")
-    @Test
-    void checkLogin() {
-        //given
-        String email = "email@email.com";
-        String password = "12345678";
-        User user = new User(new Email(email), new Password(password));
-        Mockito.when(userDao.findBy(Mockito.anyString())).thenReturn(Optional.of(user));
-        //when
-        boolean check = userService.checkLogin(email, password);
-        //then
-        assertThat(check).isTrue();
-    }
-
-    @DisplayName("이메일이 없으면 false를 반환한다.")
-    @Test
-    void checkLoginReturnFalseWithNoExistEmail() {
-        //given
-        String email = "email@email.com";
-        String password = "12345678";
-        Mockito.when(userDao.findBy(email)).thenReturn(Optional.empty());
-        //when
-        boolean check = userService.checkLogin(email, password);
-        //then
-        assertThat(check).isFalse();
-    }
-
-    @DisplayName("비밀번호가 일치하지 않으면 false를 반환한다.")
-    @Test
-    void checkLoginReturnFalseWithNotMatchesPassword() {
-        //given
-        String email = "email@email.com";
-        String password = "12345678";
-        User user = new User(new Email(email), new Password(password));
-        Mockito.when(userDao.findBy(Mockito.anyString())).thenReturn(Optional.of(user));
-        //when
-        boolean check = userService.checkLogin(email, "123456781");
-        //then
-        assertThat(check).isFalse();
     }
 }
